refactor!: remove app name in internal & docs folders
