+++ .pre-commit-config.yaml
@@ -1,4 +1,4 @@
-exclude: ^docs/lock-manager/gen/
+exclude: ^docs/gen/
 repos:
   # pre-commit install
   - repo: https://github.com/pre-commit/pre-commit-hooks
