+++ cmd/lock-manager/config.go
@@ -1,6 +1,6 @@
 package main
 
-import "github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider/repository/iredis"
+import "github.com/ruslanSorokin/lock-manager/internal/infra/provider/repository/iredis"
 
 type Config struct {
 	redis iredis.Config
