+++ docs/src/sequence/config.json
