+++ docs/src/sequence/lock.mmd
