+++ internal/infra/provider/repository/iredis/client.go
@@ -7,7 +7,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/redis/go-redis/v9"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider"
 )
 
 // Config stores all configurable parameters for redis connection.
