+++ internal/infra/provider/repository/iredis/delete_if_token_matches.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/redis/go-redis/v9"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider"
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/model"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider"
+	"github.com/ruslanSorokin/lock-manager/internal/model"
 )
 
 //nolint:gosec // no credentials
