+++ internal/infra/provider/repository/iredis/get.go
@@ -6,8 +6,8 @@ import (
 
 	"github.com/redis/go-redis/v9"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider"
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/model"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider"
+	"github.com/ruslanSorokin/lock-manager/internal/model"
 )
 
 func (s LockStorage) Get(ctx context.Context, resourceID string) (model.Lock, error) {
