+++ internal/infra/provider/repository/iredis/lock_storage.go
@@ -4,7 +4,7 @@ import (
 	"github.com/go-logr/logr"
 	"github.com/redis/go-redis/v9"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider"
 )
 
 // LockStorage stores logger & redis client.
