+++ internal/infra/provider/repository/iredis/main_test.go
@@ -13,8 +13,8 @@ import (
 	"github.com/redis/go-redis/v9"
 	"github.com/rs/zerolog"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider"
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/infra/provider/repository/iredis"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider"
+	"github.com/ruslanSorokin/lock-manager/internal/infra/provider/repository/iredis"
 )
 
 //nolint:gochecknoglobals // Using global var in tests
