+++ internal/infra/provider/test_delete.go
@@ -7,10 +7,10 @@ import (
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 
-	"github.com/ruslanSorokin/lock-manager/internal/lock-manager/model"
+	"github.com/ruslanSorokin/lock-manager/internal/model"
 )
 
-func testGet(t *testing.T, s LockProviderI) {
+func testDelete(t *testing.T, s LockProviderI) {
 	assert := assert.New(t)
 	require := require.New(t)
 
@@ -35,20 +35,23 @@ func testGet(t *testing.T, s LockProviderI) {
 
 		err := s.Create(ctx, tc.l)
 		assert.NoError(err,
-			"should create lock without any error",
+			"should create the lock without any error",
 		)
 
-		l, err := s.Get(ctx, tc.l.ResourceID)
+		err = s.Delete(ctx, tc.l.ResourceID)
 		require.NoError(err,
-			"must return lock without any error",
+			"must return the lock without any error",
 		)
-		require.Equal(l, tc.l,
-			"must be the same, as it was before inserting into the storage",
+
+		_, err = s.Get(ctx, tc.l.ResourceID)
+		assert.ErrorIsf(err, ErrLockNotFound,
+			"must return %w, as we've just deleted the lock",
+			ErrLockNotFound,
 		)
 	}
 }
 
-func testGetErrLockNotFound(t *testing.T, s LockProviderI) {
+func testDeleteErrLockNotFound(t *testing.T, s LockProviderI) {
 	require := require.New(t)
 
 	tcs := []struct {
@@ -70,7 +73,7 @@ func testGetErrLockNotFound(t *testing.T, s LockProviderI) {
 	for _, tc := range tcs {
 		ctx := context.Background()
 
-		_, err := s.Get(ctx, tc.l.ResourceID)
+		err := s.Delete(ctx, tc.l.ResourceID)
 		require.ErrorIsf(err, ErrLockNotFound,
 			"must return %w, as there is no such lock in the storage",
 			ErrLockNotFound,
